pyproject.toml
[tool.poetry]
name = "codebase-understanding"
version = "0.1.0"
description = ""
authors = ["ChihYu Yeh <chihyu.jimmy.yeh@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
tree-sitter = "^0.23.2"
tree-sitter-python = "^0.23.4"
qdrant-client = "^1.12.1"
sf-hamilton = {extras = ["sdk", "ui"], version = "^1.83.3"}
python-dotenv = "^1.0.1"
haystack-ai = "^2.7.0"
qdrant-haystack = "^7.0.0"
langfuse = "^2.56.0"
openai = "^1.57.0"
orjson = "^3.10.12"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
